# api/openapi.yaml
openapi: 3.0.3
info:
  title: Furnace Trace 入炉追溯 API
  version: 0.7.2
  description: |
    公共 API 由 core 提供，除登录/注册外需要 HttpOnly Cookie 会话。
    内部路径的 x-service 标识实际提供该接口的独立服务，使用服务 Bearer Token。
    普通用户共享只读生产视图，只能管理自己的 Telegram 绑定、偏好和投递记录。
    补录、复位、报警确认与 FUXA 网页入口要求 admin 角色。时间均为 UTC RFC3339。
    状态变更接口要求 Idempotency-Key：同一路径、同一键、同一 JSON 内容返回原响应；
    同一键不同内容返回 409。数据库唯一约束保证并发请求只有一个生效。
    通用幂等键还包含当前用户作用域。登录/注册、偏好或温度相关 PUT、解绑及配对接口不使用此键；
    配对码按单次原子消费处理，同一用户对仍存在的同一绑定重试返回当前绑定。
    已提交的幂等记录保留到显式重置演示数据；未完成记录不得按超时重新执行。
    202 只表示已接收，不能解释为设备执行完成。设备完成必须匹配 commandId 与 bootId。
    core 不访问 MQTT，所有设备状态和控制必须经 FUXA 与 scada-adapter。
    扫码/MES 拒绝或流程异常后自动排队关门、复位，匹配设备反馈后以 ABORTED 归档并恢复持续扫码；
    无需人工点击复位。设备尚未恢复时自动等待和重试关闭/复位，绝不补发新的开门命令。
servers:
  - url: /
security:
  - SessionCookie: []
...
paths:
...
  /api/v1/furnaces:
    get:
      tags: [Public]
      operationId: listFurnaces
      responses:
        '200':
          description: 全部已配置炉子及实时状态
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Furnace' }
        '503': { $ref: '#/components/responses/Error' }
  /api/v1/furnaces/{furnaceId}:
    parameters: [{ $ref: '#/components/parameters/FurnaceId' }]
    get:
      tags: [Public]
      operationId: getFurnace
      responses:
        '200':
          description: 单台炉子的可恢复快照
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Furnace' }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/furnaces/{furnaceId}/manual-scans:
    parameters: [{ $ref: '#/components/parameters/FurnaceId' }]
    post:
      tags: [Public]
      operationId: manualScan
      summary: 管理员对扫码失败的周期补录；始终执行 MES 校验，操作人取登录身份
      x-required-role: admin
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ManualScan' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /api/v1/furnaces/{furnaceId}/resets:
    parameters: [{ $ref: '#/components/parameters/FurnaceId' }]
    post:
      tags: [Public]
      operationId: resetFurnace
      summary: 管理员备用复位接口；正常界面由系统自动复位，无需点击
      description: 请求立即入队；系统最多等待十五秒取得新鲜炉门状态并等待旧命令到期，随后自动关门、复位，以设备匹配反馈中止周期。不要求入口传感器为空或其他设备在线，不将未完成周期记为成功。
      x-required-role: admin
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperatorAction' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/cycles:
    get:
      tags: [Public]
      operationId: listCycles
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: furnaceId, in: query, schema: { type: string } }
        - { name: basketNo, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/CycleStatus' } }
      responses:
        '200':
          description: 按创建时间倒序分页
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CyclePage' }
        '400': { $ref: '#/components/responses/Error' }
  /api/v1/cycles/{cycleId}:
    parameters:
      - { name: cycleId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Public]
      operationId: getCycle
      responses:
        '200':
          description: 周期及按发生顺序排列的事件与命令
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CycleDetail' }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/alarms:
    get:
      tags: [Public]
      operationId: listAlarms
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: furnaceId, in: query, schema: { type: string } }
        - { name: active, in: query, schema: { type: boolean } }
      responses:
        '200':
          description: 活动及历史报警；确认和恢复为不同字段
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlarmPage' }
        '400': { $ref: '#/components/responses/Error' }
  /api/v1/alarms/{alarmId}/acknowledgements:
    parameters:
      - { name: alarmId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [Public]
      operationId: acknowledgeAlarm
      summary: 管理员确认已查看，不清除故障，不放行设备
      x-required-role: admin
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperatorAction' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/notifications:
    get:
      tags: [Public]
      operationId: listNotifications
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: 仅当前用户的通知投递状态；unknown 表示外部发送结果不确定
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationPage' }
        '503': { $ref: '#/components/responses/Error' }
...
  /internal/v1/device-states:
    post:
      tags: [CoreInternal]
      x-service: core
      operationId: ingestDeviceStates
      security: [{ ServiceToken: [] }]
      summary: 接收适配器转发的状态；以 deviceId/bootId/seq 去重，离线状态允许重复
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              items: { $ref: '#/components/schemas/DeviceView' }
      responses:
        '204': { description: 状态已接收；不是业务完成确认 }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /internal/v1/scan-results:
    post:
      tags: [CoreInternal]
      x-service: core
      operationId: ingestScanResult
      security: [{ ServiceToken: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScanResult' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    ServiceToken: { type: http, scheme: bearer }
    SessionCookie: { type: apiKey, in: cookie, name: iot_session }
  parameters:
...
    FurnaceId:
      name: furnaceId
      in: path
      required: true
      schema: { type: string, pattern: '^[a-zA-Z0-9_-]{1,64}$' }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
  responses:
...
    Error:
      description: 结构化错误；message 不包含凭据或内部堆栈
      content:
        application/json:
          schema: { $ref: '#/components/schemas/APIError' }
    Accepted:
      description: 请求已接收；重复请求返回同一对象，不代表设备执行完成
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Accepted' }
  schemas:
...
    AdmissionSnapshot:
      type: object
      additionalProperties: false
      required:
      - basketNo
      - basketVersion
      - basketEnabled
      - batchId
      - batchNo
      - batchVersion
      - batchEnabled
      - materialName
      - processSpec
      - quantity
      - qualityStatus
      - checkedAt
      properties:
        basketNo:
          type: string
          maxLength: 64
        basketVersion:
          type: integer
          format: int64
          minimum: 1
        basketEnabled:
          type: boolean
        batchId:
          type: string
          format: uuid
        batchNo:
          type: string
          maxLength: 40
        batchVersion:
          type: integer
          format: int64
          minimum: 1
        batchEnabled:
          type: boolean
        materialName:
          type: string
          maxLength: 100
        processSpec:
          type: string
          maxLength: 200
        quantity:
          type: integer
        qualityStatus:
          type: string
          enum:
          - pending
          - passed
          - failed
        checkedAt:
          type: string
          format: date-time
...
    NotificationCategory:
      type: string
      enum: [scan, mes, device, workflow, temperature, rules]
...
    APIError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
    Accepted:
      type: object
      required: [id, status]
      properties:
        id: { type: string }
        status: { type: string, enum: [accepted, ignored] }
    CycleStatus:
      type: string
      enum: [SCANNING, VALIDATING, OPENING, WAITING_ENTRY, CLOSING, RESETTING, NEEDS_INPUT, BLOCKED, COMPLETED, ABORTED]
    Cycle:
      type: object
      required: [id, furnaceId, basketNo, rawCode, status, reason, createdAt, updatedAt]
      properties:
        mes:
          nullable: true
          allOf: [{ $ref: '#/components/schemas/MESResult' }]
        id: { type: string, format: uuid }
        furnaceId: { type: string }
        basketNo: { type: string }
        rawCode: { type: string }
        status: { $ref: '#/components/schemas/CycleStatus' }
        reason: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
    CycleEvent:
      type: object
      required: [id, cycleId, kind, message, at]
      properties:
        id: { type: string, format: uuid }
        cycleId: { type: string, format: uuid }
        kind: { type: string }
        message: { type: string }
        at: { type: string, format: date-time }
    CycleDetail:
      type: object
      required: [cycle, events, commands]
      properties:
        cycle: { $ref: '#/components/schemas/Cycle' }
        events: { type: array, items: { $ref: '#/components/schemas/CycleEvent' } }
        commands: { type: array, items: { $ref: '#/components/schemas/CommandRecord' } }
    DeviceState:
      type: object
      description: MQTT state 原文；values 字段按 kind 使用 present/inFurnace、doorOpen/doorClosed、lampOn、temperatureC（摄氏度），以及 executionCount。
      required: [deviceId, furnaceId, kind, bootId, seq, observedAt, values]
      properties:
        deviceId: { type: string }
        furnaceId: { type: string }
        kind: { type: string, enum: [sensor, door, lamp, temperature] }
        bootId: { type: string, format: uuid }
        seq: { type: integer, format: int64, minimum: 1 }
        observedAt: { type: string, format: date-time }
        values: { type: object, additionalProperties: true }
        commandId: { type: string }
        executionStatus: { type: string, enum: [IDLE, ACCEPTED, COMPLETED, REJECTED] }
    DeviceView:
      type: object
      required: [deviceId, furnaceId, kind, online, reason, state]
      properties:
        deviceId: { type: string }
        furnaceId: { type: string }
        kind: { type: string, enum: [sensor, door, lamp, temperature] }
        online: { type: boolean }
        reason: { type: string }
        state:
          allOf: [{ $ref: '#/components/schemas/DeviceState' }]
          nullable: true
    DeviceCommand:
      oneOf:
        - properties: { action: { enum: [SET_TEMPERATURE] } }
          required: [temperatureC]
        - properties: { action: { enum: [OPEN, CLOSE, RESET, ALARM_ON, ALARM_OFF] } }
      description: SET_TEMPERATURE 由 FUXA 现场输入生成，只用于温度模拟器，必须提供有限的 temperatureC；保留启动编号、有效期和去重约束。
      type: object
      required: [commandId, cycleId, deviceId, furnaceId, targetBootId, action, issuedAt, expiresAt]
      properties:
        commandId: { type: string, format: uuid }
        cycleId: { type: string }
        deviceId: { type: string }
        furnaceId: { type: string }
        temperatureC: { type: number, format: double, description: SET_TEMPERATURE 必填，摄氏度 }
        targetBootId: { type: string, format: uuid }
        action: { type: string, enum: [OPEN, CLOSE, RESET, ALARM_ON, ALARM_OFF, SET_TEMPERATURE] }
        issuedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
    CommandRecord:
      type: object
      required: [command, status, attempts]
      properties:
        command: { $ref: '#/components/schemas/DeviceCommand' }
        status: { type: string, enum: [PENDING, ACCEPTED, COMPLETED, EXPIRED, CANCELLED, UNKNOWN] }
        attempts: { type: integer, minimum: 0 }
    CameraState:
      type: object
      required: [cameraId, online, sessionId, lastFrameAt]
      properties:
        cameraId: { type: string }
        online: { type: boolean }
        sessionId: { type: string }
        lastFrameAt: { type: string, format: date-time, nullable: true }
    Furnace:
      type: object
      required: [furnaceId, name, ready, phase, cycle, devices, camera, previewUrl]
      properties:
        furnaceId: { type: string }
        name: { type: string }
        ready: { type: boolean }
        phase: { type: string, description: IDLE 或 CycleStatus }
        cycle:
          allOf: [{ $ref: '#/components/schemas/Cycle' }]
          nullable: true
        devices: { type: array, items: { $ref: '#/components/schemas/DeviceView' } }
        camera: { $ref: '#/components/schemas/CameraState' }
        previewUrl: { type: string }
...
    ScanResult:
      type: object
      required: [sessionId, cycleId, furnaceId, cameraId, outcome, rawCode, observedAt]
      properties:
        sessionId: { type: string, format: uuid }
        cycleId: { type: string, format: uuid }
        furnaceId: { type: string }
        cameraId: { type: string }
        outcome: { type: string, enum: [DECODED, SCAN_TIMEOUT, CAMERA_OFFLINE, AMBIGUOUS] }
        rawCode: { type: string, maxLength: 2048 }
        observedAt: { type: string, format: date-time }
    ManualScan:
      type: object
      additionalProperties: false
      required: [cycleId, rawCode, reason]
      properties:
        cycleId: { type: string, format: uuid }
        rawCode: { type: string, minLength: 1, maxLength: 2048 }
        operator: { type: string, maxLength: 80, deprecated: true, description: 忽略此值，审计使用登录用户 }
        reason: { type: string, minLength: 1, maxLength: 500 }
    OperatorAction:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        operator: { type: string, maxLength: 80, deprecated: true, description: 忽略此值，审计使用登录用户 }
        reason: { type: string, minLength: 1, maxLength: 500 }
    Alarm:
      type: object
      required: [id, furnaceId, cycleId, code, message, raisedAt, acknowledgedAt, recoveredAt, operator]
      properties:
        category: { $ref: '#/components/schemas/NotificationCategory' }
        ruleId: { type: string, format: uuid }
        ruleVersion: { type: integer, format: int64 }
        resolution: { type: string, enum: [condition_cleared, rule_disabled, rule_deleted, rule_migrated], description: 自定义报警结束原因，停用/删除规则不等于现场故障已消除 }
        lamp: { type: boolean, description: 是否驱动现场声光，缺省为系统报警的 true }
        id: { type: string, format: uuid }
        furnaceId: { type: string }
        cycleId: { type: string }
        code: { type: string }
        message: { type: string }
        raisedAt: { type: string, format: date-time }
        acknowledgedAt: { type: string, format: date-time, nullable: true }
        recoveredAt: { type: string, format: date-time, nullable: true }
        operator: { type: string }
    NotificationRequest:
      type: object
      required: [id, alarmId, furnaceId, kind, text, createdAt]
      properties:
        id: { type: string, format: uuid }
        alarmId: { type: string, format: uuid }
        furnaceId: { type: string }
        kind: { type: string, enum: [ALARM, RECOVERY, REMINDER] }
        category: { $ref: '#/components/schemas/NotificationCategory' }
        text: { type: string, minLength: 1, maxLength: 4000 }
        createdAt: { type: string, format: date-time }
    Notification:
      allOf:
        - { $ref: '#/components/schemas/NotificationRequest' }
        - type: object
          required: [status, attempts, lastError, sentAt, nextAttemptAt]
          properties:
            status: { type: string, enum: [pending, sent, failed, unknown, cancelled] }
            attempts: { type: integer, minimum: 0 }
            lastError: { type: string }
            sentAt: { type: string, format: date-time, nullable: true }
            nextAttemptAt: { type: string, format: date-time }
    MESResult:
      type: object
      required: [basketNo, result, message, reason]
      description: 从 PostgreSQL 台账校验；0 不存在，1 允许，2 禁用或质检未通过。未知编号不得放行。
      properties:
        reason: { type: string }
        snapshot:
          nullable: true
          allOf: [{ $ref: '#/components/schemas/AdmissionSnapshot' }]
        basketNo: { type: string }
        result: { type: integer, enum: [0, 1, 2] }
        message: { type: string }
    Pagination:
      type: object
      required: [page, pageSize, total]
      properties:
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
    CyclePage:
      type: object
      required: [items, pagination]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Cycle' } }
        pagination: { $ref: '#/components/schemas/Pagination' }
    AlarmPage:
      type: object
      required: [items, pagination]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Alarm' } }
        pagination: { $ref: '#/components/schemas/Pagination' }
    NotificationPage:
      type: object
      required: [items, pagination]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Notification' } }
        pagination: { $ref: '#/components/schemas/Pagination' }
...